Simplify functional tests
