--- Tests/Functional/Security/UserProviderTest.php
@@ -1,15 +1,0 @@
-<?php
-
-declare(strict_types=1);
-
-namespace SmartAssert\UsersSecurityBundle\Tests\Functional\Security;
-
-use SmartAssert\UsersSecurityBundle\Security\UserProvider;
-
-class UserProviderTest extends AbstractBaseFunctionalTest
-{
-    public function testServiceExistsInContainer(): void
-    {
-        $this->assertServiceExistsInContainer(UserProvider::class);
-    }
-}
